Rename cluster node metrics endpoint and create global metrics endpoint
