+++ graylog2-server/src/main/java/org/graylog2/rest/resources/cluster/ClusterMetricsResource.java
@@ -23,28 +23,20 @@ import io.swagger.annotations.ApiParam;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
-import org.apache.shiro.authz.annotation.RequiresPermissions;
-import org.graylog2.cluster.Node;
 import org.graylog2.cluster.NodeNotFoundException;
 import org.graylog2.cluster.NodeService;
 import org.graylog2.rest.RemoteInterfaceProvider;
 import org.graylog2.rest.models.system.metrics.requests.MetricsReadRequest;
-import org.graylog2.rest.models.system.metrics.responses.MetricNamesResponse;
 import org.graylog2.rest.models.system.metrics.responses.MetricsSummaryResponse;
 import org.graylog2.shared.rest.resources.ProxiedResource;
 import org.graylog2.shared.rest.resources.system.RemoteMetricsResource;
-import org.graylog2.shared.security.RestPermissions;
-import retrofit2.Response;
 
 import javax.inject.Inject;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
-import javax.ws.rs.GET;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
-import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
@@ -52,13 +44,12 @@ import java.io.IOException;
 import java.util.Map;
 import java.util.Optional;
 
-import static javax.ws.rs.core.Response.Status.BAD_GATEWAY;
-
 @RequiresAuthentication
 @Api(value = "Cluster/Metrics", description = "Cluster-wide Internal Graylog metrics")
-@Path("/cluster")
+@Path("/cluster/metrics")
 @Produces(MediaType.APPLICATION_JSON)
 public class ClusterMetricsResource extends ProxiedResource {
+
     @Inject
     public ClusterMetricsResource(NodeService nodeService,
                                   RemoteInterfaceProvider remoteInterfaceProvider,
@@ -66,73 +57,15 @@ public class ClusterMetricsResource extends ProxiedResource {
         super(httpHeaders, nodeService, remoteInterfaceProvider);
     }
 
-    private RemoteMetricsResource getResourceForNode(String nodeId) throws NodeNotFoundException {
-        final Node targetNode = nodeService.byNodeId(nodeId);
-        return remoteInterfaceProvider.get(targetNode, this.authenticationToken, RemoteMetricsResource.class);
-    }
-
-    @GET
-    @Timed
-    @Path("/{nodeId}/metrics/names")
-    @ApiOperation(value = "Get all metrics keys/names from node")
-    @RequiresPermissions(RestPermissions.METRICS_ALLKEYS)
-    public MetricNamesResponse metricNames(@ApiParam(name = "nodeId", value = "The id of the node whose metrics we want.", required = true)
-                                           @PathParam("nodeId") String nodeId) throws IOException, NodeNotFoundException {
-        final Response<MetricNamesResponse> result = getResourceForNode(nodeId).metricNames().execute();
-        if (result.isSuccessful()) {
-            return result.body();
-        } else {
-            throw new WebApplicationException(result.message(), BAD_GATEWAY);
-        }
-    }
-
     @POST
     @Timed
-    @Path("/{nodeId}/metrics/multiple")
-    @ApiOperation("Get the values of multiple metrics at once from node")
-    @ApiResponses(value = {
-            @ApiResponse(code = 400, message = "Malformed body")
-    })
-    public MetricsSummaryResponse multipleMetrics(@ApiParam(name = "nodeId", value = "The id of the node whose metrics we want.", required = true)
-                                                  @PathParam("nodeId") String nodeId,
-                                                  @ApiParam(name = "Requested metrics", required = true)
-                                                  @Valid @NotNull MetricsReadRequest request) throws IOException, NodeNotFoundException {
-        final Response<MetricsSummaryResponse> result = getResourceForNode(nodeId).multipleMetrics(request).execute();
-        if (result.isSuccessful()) {
-            return result.body();
-        } else {
-            throw new WebApplicationException(result.message(), BAD_GATEWAY);
-        }
-    }
-
-    @GET
-    @Timed
-    @Path("/{nodeId}/metrics/namespace/{namespace}")
-    @ApiOperation(value = "Get all metrics of a namespace from node")
-    @ApiResponses(value = {
-            @ApiResponse(code = 404, message = "No such metric namespace")
-    })
-    public MetricsSummaryResponse byNamespace(@ApiParam(name = "nodeId", value = "The id of the node whose metrics we want.", required = true)
-                                              @PathParam("nodeId") String nodeId,
-                                              @ApiParam(name = "namespace", required = true)
-                                              @PathParam("namespace") String namespace) throws IOException, NodeNotFoundException {
-        final Response<MetricsSummaryResponse> result = getResourceForNode(nodeId).byNamespace(namespace).execute();
-        if (result.isSuccessful()) {
-            return result.body();
-        } else {
-            throw new WebApplicationException(result.message(), BAD_GATEWAY);
-        }
-    }
-
-    @POST
-    @Timed
-    @Path("/metrics/multiple")
+    @Path("/multiple")
     @ApiOperation(value = "Get all metrics of all nodes in the cluster")
     @ApiResponses(value = {
             @ApiResponse(code = 400, message = "Malformed body")
     })
     public Map<String, Optional<MetricsSummaryResponse>> multipleMetricsAllNodes(@ApiParam(name = "Requested metrics", required = true)
-                                                          @Valid @NotNull MetricsReadRequest request) throws IOException, NodeNotFoundException {
+                                                                                 @Valid @NotNull MetricsReadRequest request) throws IOException, NodeNotFoundException {
         return getForAllNodes(remoteMetricsResource -> remoteMetricsResource.multipleMetrics(request), createRemoteInterfaceProvider(RemoteMetricsResource.class));
     }
 }
